requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.8.2
boto3==1.35.49
botocore==1.35.49
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.0.3
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
jmespath==1.0.1
MarkupSafe==3.0.2
mistletoe==1.4.0
numpy==2.1.2
openai==1.54.1
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
pyTelegramBotAPI==4.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.6
requests==2.32.3
rsa==4.9
s3transfer==0.10.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
telegramify-markdown==0.1.15
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.2
wheel==0.44.0
